requirements.txt
fastapi
uvicorn
python-dotenv
requests
numpy
pydantic
sentence-transformers
torch